requirements.txt
beautifulsoup4==4.14.2
cbor==1.0.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
chest==0.2.3
cryptography==46.0.2
Deprecated==1.2.18
dill==0.4.0
elastic-transport==9.2.0
elasticsearch==9.1.1
faiss-cpu==1.12.0
filelock==3.20.0
fsspec==2025.9.0
HeapDict==1.0.1
hf-xet==1.1.10
huggingface-hub==0.36.0
idna==3.11
ijson==3.4.0.post0
inscriptis==2.6.0
ir_datasets==0.5.11
ir_measures==0.4.1
Jinja2==3.1.6
joblib==1.5.2
lxml==6.0.2
lz4==4.4.4
MarkupSafe==3.0.3
more-itertools==10.8.0
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
pyarrow==21.0.0
pycparser==2.23
pyjnius==1.7.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-terrier==0.13.1
pytrec_eval-terrier==0.5.10
pytz==2025.2
PyYAML==6.0.3
regex==2025.10.23
requests==2.32.5
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
statsmodels==0.14.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
trec-car-tools==2.6
typing_extensions==4.15.0
tzdata==2025.2
unlzw3==0.2.3
urllib3==2.5.0
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
wrapt==1.17.3
zlib-state==0.1.10
